+++ lib/src/utils/share/share_post.dart
@@ -1,19 +1,20 @@
 import 'package:feed_sx/feed.dart';
 import 'package:feed_sx/src/services/likeminds_service.dart';
+import 'package:feed_sx/src/utils/credentials/credentials.dart';
 import 'package:likeminds_feed/likeminds_feed.dart';
 import 'package:share_plus/share_plus.dart';
-
-import 'package:feed_sx/src/utils/local_preference/user_local_preference.dart';
-
 part 'deep_link_request.dart';
 part 'deep_link_response.dart';
 
 class SharePost {
+  static String userId = prodFlag ? CredsProd.botId : CredsDev.botId;
+  static String apiKey = prodFlag ? CredsProd.apiKey : CredsDev.apiKey;
+  // TODO: Add domain to your application
+  String domain = 'feedsx://www.feedsx.com';
   // fetches the domain given by client at time of initialization of Feed
 
   // below function creates a link from domain and post id
   String createLink(String postId) {
-    final String domain = UserLocalPreference.instance.getAppDomain();
     int length = domain.length;
     if (domain[length - 1] == '/') {
       return "$domain/post?post_id=$postId";
@@ -44,6 +45,22 @@ class SharePost {
     List secondPathSegment = request.link.split('post_id=');
     if (secondPathSegment.length > 1 && secondPathSegment[1] != null) {
       String postId = secondPathSegment[1];
+      setupLMFeed(request.callback, request.apiKey);
+      await locator<LikeMindsService>()
+          .initiateUser((InitiateUserRequestBuilder()
+                ..apiKey(request.apiKey)
+                ..userId(request.userUniqueId)
+                ..userName(request.userName))
+              .build());
+
+      locator<NavigationService>().navigateTo(
+        AllCommentsScreen.route,
+        arguments: AllCommentsScreenArguments(
+          postId: postId,
+          feedRoomId: request.feedRoomId,
+          fromComment: false,
+        ),
+      );
       return DeepLinkResponse(
         success: true,
         postId: postId,